assignment 2modulel2

+++ app/blueprints/serviceTickets/routes.py
@@ -32,24 +32,30 @@ def edit_service_tickets(service_ticket_id):
     query =select(ServiceTickets).where(ServiceTickets.id== service_ticket_id)
     service_ticket = db.session.execute(query).scalars().first()
 
+    if not service_ticket:
+        return jsonify({'message': 'Service ticket not found'}), 404
+
+    # add mechanics
     for mechanic_id in ticket_edits['add_mechanic_ids']:
-        query = select(Mechanic).where(Mechanic.id == mechanic_id)
-        mechanic= db.session.execute(query).scalars().first()
-    if mechanic and mechanic not in service_ticket.mechanics:
-        service_ticket.mechanics.append(mechanic)
+        m_query = select(Mechanic).where(Mechanic.id == mechanic_id)
+        mechanic = db.session.execute(m_query).scalars().first()
+        if mechanic and mechanic not in service_ticket.mechanics:
+            service_ticket.mechanics.append(mechanic)
     
+    # remove mechanics
     for mechanic_id in ticket_edits['remove_mechanic_ids']:
-        query = select(Mechanic).where(Mechanic.id == mechanic_id)
-        mechanic= db.session.execute(query).scalars().first()
-    if mechanic and mechanic  in service_ticket.mechanics:
-        service_ticket.mechanics.remove(mechanic) 
+        m_query = select(Mechanic).where(Mechanic.id == mechanic_id)
+        mechanic = db.session.execute(m_query).scalars().first()
+        if mechanic and mechanic in service_ticket.mechanics:
+            service_ticket.mechanics.remove(mechanic)
     db.session.commit()
     return return_ticket_schema.jsonify(service_ticket),200
 
-@service_tickets_bp.route('/<int:service_ticket_id>', methods=['PUT'])
-def delete_tiocket(ticket_id):
-    query=select(ServiceTickets).where(ServiceTickets.id== service_ticket_id)
-    service_ticket = db.session.execute(query).scalars().first()
+@service_tickets_bp.route('/<int:service_ticket_id>', methods=['DELETE'])
+def delete_ticket(service_ticket_id):
+    service_ticket = db.session.get(ServiceTickets, service_ticket_id)
+    if not service_ticket:
+        return jsonify({'message': 'Service ticket not found'}), 404
     db.session.delete(service_ticket)
     db.session.commit()
     return jsonify({'message': f'Service ticket id:{service_ticket_id} was deleted successfully'}), 200
